refactor(resolver): extract fs/env helpers to keep resolution.rs under ratchet
Adding the realpath memoization (`canonical_dir`) pushed `resolution.rs`
over its architecture size ratchet (319 > 301). Move the self-contained
filesystem/environment helpers (`is_declaration_file`,
`canonicalize_with_missing_tail`, `canonicalize_or_owned`, `env_flag`) into a
new `resolution/fs_helpers.rs` module and re-export them, keeping every
caller's `crate::driver::resolution::<item>` import path stable. The
orchestrator drops to 280 lines; tighten the ratchet to 280.
